Test client seems to be working...
Still, I'll try to create a "test_drm" program that also initialize
a CRTC with an encoder, map the framebuffer and display something
on the screen by writing pixels on the buffer, to be sure that dumb
buffers are really usable.

Once done, we'll initialize the VPU driver phase where we'll receive
the FD and try to map the buffer to the VPU this time !

Then, whether it works or not, I'll try to contact some IOMMU, DMA and
DRM knowledgeable people, see if they can help me correct the VPU
driver part.

And then, I'll try to reimport everything back in the real VPU driver.

Signed-off-by: Myy <myy@miouyouyou.fr>