Now printing the resources defined on the VPU
And added a few helpers functions for that.

Now, the point is to understand how to write something in the VPU
memory using its IOMMU. Once done, we can try to perform the DMA
operation from the GPU to the VPU and, then, check if the data got
copied correctly.

Signed-off-by: Myy <myy@miouyouyou.fr>